QResampler, rewrite of RationalFactorResampler.
This is a rewrite of the C++ RationalFactorResampler resampling library.
The goals are to make it easier to use ProcessSamples() and Flush() in a
way that doesn't cause heap allocations, add multichannel support, and
to simplify the interface. The top-level of comment of
audio/dsp/resampler_q.h has a QResampler user guide and further details.

WARNING: RationalFactorResampler is DEPRECATED in this commit and will be
removed in a following commit.

Users of RationalFactorResampler should migrate to QResampler.
QResampler produces the same resampled output as RationalFactorResampler
to near machine precision, so this change should have negligible effect
on how signals are resampled. Please see the migration notes at the top
of audio/dsp/resampler_rational_factor.h.